% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.r
\name{feed}
\alias{feed}
\title{Feed Argos track(s) to ssmTMB}
\usage{
feed(dat, dlen = NULL, ts = 2, pass2 = FALSE, plot = FALSE, ...)
}
\arguments{
\item{dat}{A data frame containing the following columns, "id","date",
"lc", "lon", "lat". "id" is a unique identifier for the tracking dataset.
"date" is the GMT date-time of each observation with the following format
"2001-11-13 07:59:59". "lc" is the Argos location quality class of each
observation, values in ascending order of quality are "Z", "B", "A", "0", "1",
"2", "3". "lon" is the observed longitude in decimal degrees. "lat" is the
observed latitude in decimal degress. The Z-class locations are assumed to 
have the same error distributions as B-class locations.}

\item{dlen}{index to trim input data down to dlen tracks for testing}

\item{ts}{time step in h}

\item{pass2}{logical, try a second optimisation using optim (much slower than nlminb)}

\item{plot}{logical, generate plot of fits to data. File name is generated from name of}

\item{...}{other arguments to \code{ssmTMB::fit_ssm}}

\item{span}{parameter that controls the degree of smoothing by \code{stats::loess},
used to obtain initial values for the location states. Smaller values = less
smoothing. Values > 0.3 may be required for sparse datasets}
}
\value{
A list is returned with each outer list elements corresponding to each unique 
individual id in the input data.
}
\description{
This is a wrapper function that feeds single or multiple Argos datasets for filtering via
\code{ssmTMB::fit_ssm}
}
\examples{
\dontrun{
}
}
\author{
Ian Jonsen
}

