% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtrim.r
\name{gtrim}
\alias{gtrim}
\title{Trims deployment start and end periods of spurious data using a 
  (primitive) graphical interface}
\usage{
gtrim(dat)
}
\arguments{
\item{dat}{track data that has already been \code{strip}'ped}
}
\description{
Use graphical interface to identify proper start and end of animal tracks
}
\examples{
\dontrun{
}
}
\author{
Ian Jonsen
}

