% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fit.r
\name{plot_fit}
\alias{plot_fit}
\title{Visual an SSM fit to data as a map and as 1-d time-series}
\usage{
plot_fit(ssm, dat = NULL, sp, dlen = NULL, res = "high")
}
\description{
Visual an SSM fit to data as a map and as 1-d time-series
}

