% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotf.r
\name{plotf}
\alias{plotf}
\title{Visual an \code{ssmTMB} fit to data as a map and as 1-d time-series}
\usage{
plotf(ssm, res)
}
\description{
Visual an \code{ssmTMB} fit to data as a map and as 1-d time-series
}

