% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plots}
\alias{plots}
\title{Visualise a \code{ssmTMB} fit to data as a map and as 1-d time-series}
\usage{
plots(ssm, d, res)
}
\description{
Visualise a \code{ssmTMB} fit to data as a map and as 1-d time-series
}

