% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefilter.r
\name{prefilter}
\alias{prefilter}
\title{Prefilters extreme outlier locations in Argos data}
\usage{
prefilter(d, vmax = 10, w = 5)
}
\arguments{
\item{d}{A data_frame containing the following columns: 
"id", "date", "lc", "lon", "lat". "id" is a unique identifier for the tracking dataset.
"date" is the GMT date-time of each observation with the following format
"2001-11-13 07:59:59". "lc" is the Argos location quality class of each
observation, values in ascending order of quality are "Z", "B", "A", "0", "1",
"2", "3". "lon" is the observed longitude in decimal degrees. "lat" is the
observed latitude in decimal degress.}
}
\value{
A tbl_df grouped by individual id is returned.
}
\description{
Use \code{diveMove::grpSpeedFilter} to identify extreme 
outlier locations prior to \code{feed}-ing to \code{ssmTMB::fit_ssm}. Outliers are not 
removed but flagged with a logical vector so they can be ignored when \code{feed}-ing 
to \code{ssmTMB::fit_ssm}. \code{prefilter} also ensures data are temporally ordered.
}
\examples{
\dontrun{
}
}
\author{
Ian Jonsen
}

