% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfilter.r
\name{sfilter}
\alias{sfilter}
\title{State-space Filter Argos track(s) via \code{ssmTMB::fit_ssm}}
\usage{
sfilter(d, ts = 4, ...)
}
\arguments{
\item{d}{data output from prefilter() or trim(), ie. a tibble data_frame, grouped by individual id
with a 'keep' logical variable indicating which observations are to be ignored}

\item{ts}{time step in hours}

\item{...}{other arguments to \code{ssmTMB::fit_ssm}}
}
\value{
A list is returned with each outer list elements corresponding to each unique 
individual id in the input data.
}
\description{
This is a wrapper function that feeds single or multiple Argos datasets for state-space filtering via
\code{ssmTMB::fit_ssm}
}
\examples{
\dontrun{
}
}
\author{
Ian Jonsen
}

