% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip.r
\name{strip}
\alias{strip}
\title{strip out extreme outlier locations from Argos data}
\usage{
strip(dat, min.obs = 20, min.days = 1, vmax = 10, dmax = 500)
}
\arguments{
\item{dat}{A data_frame containing the following columns: 
"id", "date", "lc", "lon", "lat". "id" is a unique identifier for the tracking dataset.
"date" is the GMT date-time of each observation with the following format
"2001-11-13 07:59:59". "lc" is the Argos location quality class of each
observation, values in ascending order of quality are "Z", "B", "A", "0", "1",
"2", "3". "lon" is the observed longitude in decimal degrees. "lat" is the
observed latitude in decimal degress.}
}
\value{
A list is returned with each outer list element corresponding to each unique 
individual id in the input data. The output data_frames include an additional column, labelled
"filt" which is a logical vector indicating whether location is to be used in subsequent steps
}
\description{
Use \code{diveMove::grpSpeedFilter} & \code{geosphere::distGeo} to identify extreme 
outlier locations prior to \code{feed}-ing to \code{ssmTMB::fit_ssm}. Outliers are not 
removed but flagged with a logical vector so they can be ignored when \code{feed}-ing 
to \code{ssmTMB::fit_ssm}. \code{geosphere::distGeo} is only used for the first and
last locations, which aren't handled by \code{diveMove::grpSpeedFilter}.
}
\examples{
\dontrun{
}
}
\author{
Ian Jonsen
}

