% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip.r
\name{strip}
\alias{strip}
\title{Strip Argos track data of unwanted records}
\usage{
strip(dat, what = rep(TRUE, 6), min.obs = 30, min.days = 10)
}
\arguments{
\item{dat}{A data.frame containing the following columns:
"id", "date", "lc", "lon", "lat". "id" is a unique identifier for the tracking dataset.
"date" is the GMT date-time of each observation with the following format
"2001-11-13 07:59:59". "lc" is the Argos location quality class of each
observation, values in ascending order of quality are "Z", "B", "A", "0", "1",
"2", "3". "lon" is the observed longitude in decimal degrees. "lat" is the
observed latitude in decimal degress.}
}
\value{
A list is returned with each outer list element corresponding to each unique
individual id in the input data.
}
\description{
Strip does the following:  1) removes duplicated date records; 2) removes start locations in
N Hemisphere, eg. Seattle, BAS, SMRU (will be generalised later); 3) removes deployments with
less than min.obs records; 4) removes deployments that last less than min.days; 5) removes
records with NA lat and/or lon; 6) shifts 0, 360 longitudes to -180, 180. Each of these steps
can be optionally turned off.
}
\examples{
\dontrun{
}
}
\author{
Ian Jonsen
}

