% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip.r
\name{strip}
\alias{strip}
\title{Strip Argos track data of unwanted records}
\usage{
strip(d, what = rep(TRUE, 6), min_obs = 30, min_days = 10)
}
\arguments{
\item{d}{A data.frame containing the following columns:
"id", "date", "lc", "lon", "lat". "id" is a unique identifier for the tracking dataset.
"date" is the GMT date-time of each observation with the following format
"2001-11-13 07:59:59". "lc" is the Argos location quality class of each
observation, values in ascending order of quality are "Z", "B", "A", "0", "1",
"2", "3". "lon" is the observed longitude in decimal degrees. "lat" is the
observed latitude in decimal degress.}

\item{what}{A logical vecotr of length 6 that turns on or off each of the 6 data stripping
steps.}

\item{min_obs}{The minimum number of observations an individual track requires to be retained.}

\item{min_days}{The minimum number of deployment days an individual tracks requires to be retained.}
}
\value{
A tbl_df grouped by individual id is returned.
}
\description{
Strip does the following:  1) removes duplicated date records; 2) removes start locations in
N Hemisphere, eg. Seattle, BAS, SMRU (will be generalised later); 3) removes deployments with
less than min.obs records; 4) removes deployments that last less than min.days; 5) removes
records with NA lat and/or lon; 6) removes Z-class locations.
Each of these steps can be optionally turned off via the \code{what} logical vector.
}
\examples{
\dontrun{
}
}
\author{
Ian Jonsen
}

