% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.r
\name{trim}
\alias{trim}
\title{Trims deployment start and end periods of spurious data via
   changepoint estimation}
\usage{
trim(d, n = 11)
}
\arguments{
\item{d}{track data that has already been \code{strip}'ped}

\item{n}{window size to be used to obtain moving average of distance
from deployment (first) location}
}
\description{
Trims deployment start and end periods of spurious data via
   changepoint estimation
}
\examples{
\dontrun{
}
}
\author{
Ian Jonsen
}

